Refactor pacman block to use features of std
Calling out to a shell is not necessary for a bulk of the commands in
`get_update_count`, it probably is bad practice to do so. This commit
cleans `get_update_count` up and removes most of the shell-calls.

In addition, `run_command` has been changed to wait for the executed
commands to exit. `Command::spawn` will spawn the processes without
blocking, i.e. the program will continue with it's normal execution
immediately, even though the desired result (e.g. the update of the
pacman database) is not even completed yet.

The `trap` has been removed, since it either wouldn't have worked
reliably (since the launched process was not kept track of) or it exited
immediately, causing the `rm` call to the pacman lock to execute and
delete any lock before the other processes are done.

Deleting the pacman lock in case of a failure is a benevolent idea, but
would require full error tracking for the pacman related processes.

This should fix issue #30.