sound: Add config show_volume_when_muted
The sound block doesn't show the volume when it's muted. This commit
adds a config option to enable just that.

It is useful to see how high the volume is turned up before unmuting.
This also allows you to adjust the volume while muted and then unmute at
the right volume.