Makes it possible to read themes from standalone files.
This introduces a new `theme.file` key to the TOML configuration. We
look in the following locations for this file:

- The actual file location, if it is a real path.
- $XDG_CONFIG_HOME/i3status-rs/themes; and
- /usr/share/i3status-rs/themes

The theme file must be "complete", e.g. contain all of the keys, for it
to be serialized correctly.

(cherry picked from commit f36e2e3ae0c3b82f21ea6e3212938c5d9824a816)