Ignore stdin and stdout for `spawn_child_async`
If stdout from child processes is not deliberately ignored and/or
quieted via external means, it can interfere with the i3bar protocol
stream. Stdin is less dangerous, but gets closed almost immediately by
`wait()`ing.

Connect both of them to `Stdio::null()` to avoid any problems. Stderr
is safe and can remain as-is for debugging purposes.