fix builds on armv7 and i686
These were broken by #449, which naively fixed armv6 builds.
This commit should now work on all platforms.
It wraps all return values of statvfs from the nix crate in
`as u64`, which means that the platform specific data type
(which might be u32 instead) is always converted to u64 when
necessary. Before this, only values which were known to cause
an issue were converted.