Change length calc to better deal with multi-byte Unicode chars
Also update min_width calculation as it should be config.max_width + len(icon) + len(trailing space)

i3bar protocol:
The minimum width (in pixels) of the block. If the content of the full_text key take less space than the specified min_width,
the block will be padded to the left and/or the right side, according to the align key.
This is useful when you want to prevent the whole status line to shift when value take more or less space between each iteration.
The value can also be a string. In this case, the width of the text given by min_width determines the minimum width of the block.
This is useful when you want to set a sensible minimum width regardless of which font you are using, and at what particular size.