set backlight brightness via DBus when sysfs is inaccessible
After not getting sysfs access to work, even with the udev rule
mentioned in the documentation, I investigated how other
applications/desktops are controlling the backlight brightness.

As it turns out, on systems governed by systemd-logind (or elogind), the
correct way according to the systemd developers [1] is the DBus
interface of logind.

This commit uses this method as a fallback for when we cannot open the
sysfs file. This works great on my system, but I'm not particularly
attached to this method if the maintainers find it too surprising. If
you'd rather like a configuration option to switch between two methods,
I can implement this as well.

[1] https://github.com/systemd/systemd/issues/14463#issuecomment-570190340