music block: add smart_trim config option
This allows the block to be a little smarter in how it trims text to fit within
a max width. More precisely - it will trim from both the artist and the title
in proportion to their lengths, to try and show the most information possible.

Based on work done by https://github.com/jgbyrne mentioned here:
https://github.com/greshake/i3status-rust/issues/189