Replace i3ipc-rs with swayipc-rs
This allows future work such as adding marks support to the focused_window block,
and a sway driver for the keyboard_layout block. sway IPC is built on top of i3 IPC,
so anything that works in i3 also works in sway. However with this crate we can
access sway-only parts of the IPC such as the keyboard layout. Also, the maintainer
for i3ipc-rs seems to be AWOL at the moment which is stalling #532