toggle block: change state if toggle command failed
PR648 made it so the toggle block only toggles if the respective toggle
command exited successfully. However from a user point of view there was no
feedback, so the user might think their clicks are not registering when in
they are. This PR sets the block state to Critical when a toggle command
fails in order to give visual feedback. The state is reset once clicked again
and the command succeeds, or until the next time the command_state cmd is run.