Better fix for #535 and #582
This undoes the bandaid fixes that were implemented to deal with the panics that
occur when only using blocks that have no update interval (ones that use DBus etc).

Instead of kludging a dummy update interval to those blocks, the scheduler
logic is changed so that no updates are scheduled if the block has no
update_interval value set.