Prevent bar being stopped by using signal other than SIGSTOP
i3 sends SIGSTOP to the bar when it is hidden, which causes the bar to stop
updating. This is done for "power savings" and is not configurable in i3 at
the moment, which inconveniences users of blocks in i3status-rs that are
meant to be running at all times.

Luckily i3 lets us customise the signal it sends, so we can set it to something
other than SIGSTOP to effectively allow non-interrupted of the bar in all
situations.

Fixes #503, #694 and possibly others.