Add warning_updates_regex to pacman block
This does pretty much what you expect: allows you to specify a warning
regex using the same syntax as the critical regex, and if any packages
in the warning regex have updates, the block will have the warning style
applied to it. If a critical regex package matches, the block will have
the critical formatting applied.

The critical regex will take priority over the warning; if both regexes
match, the block will be styled as critical. There are optimizations to
the code that could be made to short-circuit checking the warning regex
if the critical regex matches, but I think that's better left to a
different PR.