Add support for update signals to custom blocks
Allows custom blocks to update if a specified SIGRT signal is recieved,
also adds support for potential future SIGRT support for other blocks.

Spawns a new thread that listens for specified signals, this uses the
signal_hook crate.

Expands the blocks trait to allow for a signal event as well as to be
able to ask it for which potential signal a block is listening to.
The get_signal() function might be unneeded and could be removed if we
are willing to listen to any RTSIG signal rather than only the ones
specified.

Currently i3status-rs will crash if it recieves a RTSIG signal that it
has not specifically asked to listen to, this was done in order to make
debugging this easier.