Restructure signal and remove selective signals
Restructure and move functions that deal with signals into their own
file.
Remove the functionality that listened only to the relevant signals
rather than every SIGRT event. The reasoning behind this is that it
required a significant amount of general code for example an extra trait
function in the blocks trait without bringing any important
functionality.
The original reason for selective signal listening was in
order to make debugging easier when a non-listened to signal caused a
crash. Now we avoid crashing instead.

Right now we are listening to every SIGRT event within the allowed span
(SIG 34 - 64). Of course blocks will only update if their corresponding
signal is sent still.