nvidia block: if label is provided then show it first
If someone sets `label` in their config then it is a pretty good bet that
they want the bar to display that instead of the default GPU name from `nvidia-smi`.
Up until now those users would have had to click the bar after startup to get the
widget to show the label, but now it will show the label first by default.

Should resolve the usecase desired by #754