backlight: support nonlinear brightness
The raw brightness values are quite useless on some hardware. Add an
option to scale the raw backlight PWM values on a power scale.

Since there is no single correct exponent even in the CIELAB L* function
and the mapping between raw hardware brightness value and relative
luminance is unknown, the implementation allows user to fine-tune the
exponent.