ibus block: allow block to initialise before IBus is up
If i3bar/swaybar was started before IBus was up then the ibus block would crash
on startup since it would either find no socket files in the ibus dir, or it
would find a stale one from last time IBus was running and then error out when
attempting to connect to the bus.

With this change we monitor D-Bus and only start the monitoring thread after
we find out that IBus is running.

TODO: With this PR the ibus block will no longer crash on startup, however if
IBus itself is restarted while the block is running it has no way of restarting
the monitoring thread with the new ibus address, so in that case the bar must be
restarted (IBus creates a different socket/address each time it is started).