music block: completely overhaul update mechanism
Currently the block subscribes to MPRIS PropertiesChanged signals and then
schedules an update unconditionally whenever it received a signal. Inside the
update() function, multiple calls are made to retrieve the Metadata,
PlaybackStatus or even to grab a new player. Not only are all these calls
redundant, the block is also unable to keep track of players which disappear
from the bus without emitting a PropertiesChanged signal (for example mpv).
In addition, there are often signals sent in which the Metadata/PlaybackStatus
has not actually changed, yet we schedule an update regardless.

Changes with this PR:
- each PropertiesChanged signal actually contains the new values of the
  properties that were changed, so we will grab them from here instead of
  making another method call.

- in another thread we will monitor the bus for new or disconnected players
  which will allow us to clear the block when a player has gone

- change our data structure so that we keep track of all current players,
  which with further work in the future should allow us to cycle between
  players by right clicking (for example) on the bar