tree wide: use simple 16 byte random string for UUID
Since all we are using UUIDs for is to avoid duplicate names then we can
achieve the same thing by using a random 16 byte string rather than strictly
adhering to the UUID v4 spec (which just sets a couple of the bytes to a
known value for identification purposes). This way we can reduce dependencies
by just relying on the bare minimum of `getrandom`.