net block: do not error out when arrays are empty
"Malformed JSON" doesn't make sense as an error since if the JSON itself was
invalid then serde would catch it earlier. We will trust that `ip` is outputting
a valid response, and if that response contains no ip address info then we can
just return None instead of erroring out.