net block: change icon for loopback device
When `device` is not set in the config, the block will automatically
grab the first device it finds in the output of `ip route show default`.
When there are no devices here it will fallback to the loopback device.
Currently the loopback device shows up as a connected ethernet device
which can be misleading, so here we special case the loopback device so
that there is no confusion.

In the future, perhaps the fallback to the loopback device should be
removed, however it will require some restructing of the net block code
so we will go with this for now.