Modify taskwarrior block to support multiple named filters
This commit deprecates `filter_tags` in favour of `filters`.
If `filter_tags` are specified, the block behaves like the
the existing implementation, and two filters are created:

- all: `-COMPLETED -DELETED`
- filtered: `-COMPLETED -DELETED <tags>`

This emulates the modes of `TaskwarriorBlockMode` which has also been
removed.

Clicking now cycles through the filters, of which an arbitrary number
may be specified.

[[block.filters]]
name = "today"
filter = "+PENDING +OVERDUE or +DUETODAY"