Refactored weather block to use curl crate
This commit makes the weather block use the curl bindings through the
`curl` crate rather than start a sub process to execute the curl binary.

This has a few advantages:

- Is faster and uses less resources as no shell is started

- We can use simpler code to parse the errors, status codes and output

- The user does not need to depend on the `curl` binary.  The `curl`
crate depends on `curl-sys` which can be configured to link to libcurl statically or dynamically.

I believe other blocks should use these bindings as well, instead of the `curl` binary.