Refactor functions out of pattern matching in weather block
Extract some functions out of the `update` function. This helps with
readability and simplifies the code a bit.

Do not depend on the state of `weather_keys` to control whether an
exception is presented to the user. Instead, use the `Err` value
returned in `update`. This is clearer to understand when reading this code.

Show better error messages when some error occurs.