Do not crash entire bar when weather block fails
Some cases were being handled, and transformed into a `x` string to be
shown to the user, but in many cases the entire bar would just crash
when an error ocurred while getting the weather.

This commit handles all `Error` on `update` and shows an error message
on the bar, instead of crashing the entire bar.