net: Add spacing due to the decimal sign
When we transition from a speed like 918KB to 2.16MB, the number of
digits remains the same, but the total length of the string is
different.

This causes all elements to the left of this block to shift, which is an
annoying visual effect.

This patch checks if a `.` character was added (by checking if decimals
> 0), if not, then we prepend a string to compensate for that.