Include "Power Draw" option for Nvidia GPU block
This fits in with the existing `nvidia-smi` invocation.

I fit this at end of the Nvidia block, as it seemed visually appealing
there (this did mean some `count++` and `.clone()` lines in the
previously-trailing option, hopefully it makes sense and is easy to
follow).

I was not able to find the exact name `power.draw` from the `nvidia-smi`
manual page, but from this site:

https://briot-jerome.developpez.com/fichiers/blog/nvidia-smi/list.txt

I found the following line that the author has documented as:

```
"power.draw"
The last measured power draw for the entire board, in watts. Only available if power management is supported. This reading is accurate to within +/- 5 watts.
```

I was able to run with this query option on my machine, as seen here:

```
 $ nvidia-smi -i 0 --query-gpu="name,memory.used,memory.total,utilization.gpu,temperature.gpu,fan.speed,clocks.current.graphics,power.draw" --format=csv,noheader
GeForce GTX 980 Ti, 836 MiB, 6080 MiB, 21 %, 62, 7 %, 683 MHz, 40.47 W
```

or more simplified:

```
 $ nvidia-smi -i 0 --query-gpu="power.draw" --format=csv,noheader
40.47 W
```

```
 $ nvidia-smi -i 0 --query-gpu="power.draw" --format=csv,noheader,nounits
40.67
```