bluetooth: allow controller to be specified
If there are multiple controllers available, and the device we are interested in
has been paired with both in the past, or otherwise shows up when we enumerate
all bluetooth devices on the system, then there is a chance we select the wrong
one, due to how the block works.

We can avoid this by allowing the controller to be specified in the config.

This also applies in the case where the bluetooth module is disabled on startup,
since we currently set a manual fallback path with a hardcoded controller id (hci0),
but this might not always be the desired controller.

It would have been neat to be able to specify the controller via its MAC address,
but the way the block works now, it requires the id (hci0, hci1, etc), and as far
as I know, it's not possible to know this beforehand in all cases.