Remove redundent set_spacing calls.
`nvidia_gpu` chooses between `Hidden` and `Inline` spacing right
beofre calling set_text, so the set_spacing is a noop.
Right before this though, in the `NameWidgetMode::ShowDefaultName`
section of the match, there is an instance of set_spacing called
after set_text is called. This is not a noop because the `set_text(s)`
calls will only set spacing to `Hidden` or `Normal`.

`sounds.rs` called `set_spacing` right before `set_texts` so again
this is a noop.

Lastly in the `temperature.rs` there were two calls to `set_spacing`
after `set_text(s)` is called, however the spacing that is set matches
the spacing that will be set by `set_text(s)` so this too is a noop.