Fix for batteries allow_missing
ammgws pointed out that UPower devices were thowing an error if the
device was not present when the bar started. If the battery has
`allow_missing=true` this is not a case that we want to throw an
error. There was a similar issue present in the new apcaccess
battery driver that is fixed as well. Both drivers are able to
pull information from a device is that not present when the status
bar is launched, when `allow_missing=true`.