Add wl-gammarelay driver for hueshift
Add support for wl-gammarelay.
While reviewing hueshift.rs It was noted that `Hueshift::update`
does not need to be called on the `update_interval`.

If drivers have a way of polling for the current temperature then it
makes sense to have an update interval otherwise it has no effect.
None of the drivers besides WlGammarelay has a mechanism to get the
current temperature if they are changed outside of the statusbar.
Although WlGammarelay can get the current temperature it doesn't need
to run update on an update interval as it is listening to dbus events.
Something like this:
```
Ok(match self.hue_shifter {
    HueShifter::X | HueShifter::Y => Some(self.update_interval.into()),
    _ => None,
})
```

Add `wl-gammarelay-rs` and `wl-gammarelay` drivers to docs.

As suggested by @MaxVerevkin:

We don't need to check to see if the first arg is `rs.wl.gammarelay`.
Instead we can just make sure `arg0namespace='rs.wl.gammarelay'` .

Check if either `wl-gammarelay-rs` or `wl-gammarelay` is installed.

Start `wl-gammarelay-rs` or `wl-gammarelay` as required.

Use `Arc<AtomicU16>` to share state between spawned thread and driver.