Don't enumerate UPower devices every update (#1437)
* Don't enumerate UPower devices every update

Fixes #1434

Instead of enumerating the devices each time the block is updated
we can ask for the device enumeration only when a device is added
or removed if allow_missing is true.

* Don't enumerate UPower devices in if_available

@SilverlightningY pointed out that we were still enumerating UPower
devices in `if_available`. Since we are watching for add and removal of
devices by updating the `device_path` to be `Arc<Mutex<None>>` we'll
know if the device is still if the device is still available or not.