Allow globbing for power_supply devices
This allows specifying a wildcard pattern for devices under
/sys/class/power_supply/

It is usefull for e.g. wireless mice, that change their device
identifier upon unplugging.

E.g. my logitech mouse starts with
/sys/class/power_supply/hidpp_battery_0/
and when it's being unplugged it changes its name to
/sys/class/power_supply/hidpp_battery_1/
and then upon replugging it's now called
/sys/class/power_supply/hidpp_battery_2/
and so on.

This allows me to now just specify
hidpp_battery_* in my config file to use the first (and only)
device that matches this pattern.

It still requires a restart of i3status-rs uppon unplugging but I will
try to fix that in a later PR.