Fix a race condition with kernel driver
Since #1456 it is possible for status() and refresh_device_info() to
fail in case the device is not available.

If the device gets removed just after checking if it is available but
before checking for it's status this could crash i3status-rs.

So in order to prevent this I swapped the order so that we first check for the
device status and refresh the device info and then check if the device is available.

Now it's guaranteed that the device was available when we checked status() or
called refresh_device_info().