battery: add a special case for formatting of zero watts
When a power adapter is connected and the battery is fully charged,
there will be no current going in or out of the battery. Therefore the
power consumption is reported as 0. Before this commit,
`Value::from_float(0).watts()` produces the display string "0.0nW" which
looks weird. This change hardcodes "0W" instead.