Improvements to apc_ups driver
Reduce the amount of `if let` chaining in using `or_error`s instead.
Removed the `unwrap`s, only using `unwrap_or` and `unwrap_or_default`
now.

Created a separate `ApcConnection` struct that contains the read/write
methods to talk to the APC UPS. I figured out the reason that
the driver could not continue to load information after `apcupsd` was
restarted is that a new TcpStream needs to be established after the
server closes the connection.