weather: Move data into modules
met_no needs a one-time fetch of the "legend", the translation table
from weather summary keywords to textual representations. This is now
stored in the Config struct within the module, and fetched on first get.

Adds a trait WeatherProvider with a single function get_weather that
takes a mutable self, to allow for this update.

The main loop instantiates the config within an Arc<Mutex>>. Async
FnMut enclosures cannot have references that outlive the enclosure, so
to use api.recoverable we need to do an Arc clone and lock. This also
makes sure only one request can be run at a time.

Also moves the autolocation call inside the api.recoverable block.