weather: Add support for met.no (#1577)
* weather: Add support for met.no

Refactors code for OpenWeatherMap into its own module, and adds a new
module for met.no.

* weather(met.no): fix clippy warnings

Remove a couple of unneccessary clones and some nonidiomatic code.

* fixup(weather): move modules into separate files

* fixup(weather): Avoid duplication of config blocks

Converted the enums in weather to hold an inner config block, and then
pass just that config straight to the get method.

This means the autolocation data must be passed separately to both
implementations.

* fixup(weather): Rename OPEN_WEATHER_MAP_* consts

The prefix is no longer needed when all the code is namespaced in a
module.

* fixup(weather): Move init functions to module

Move the getenv_openweathermap_* and default_lang() functions to the
open_weather_map module.

* doc(weather): Tell when coordinates are required

* fixup(weather): Remove double reference

* fixup(weather): Use or_else() instead of or()

Clippy complaint.

* fixup(weather): Default for met_no lang

* fixup(weather): Remove unused From for ApiLanguage

* fix(weather): Fetch met_no legend only once

Moved the fetching of legends into the initialization part of the run
method in weather.rs. Pass a reference to the legends hashmap to
met_no::get().

Also clean up the translate function, with a more reasonable way to
produce the default value when a legend is missing.

* fixup: use unwrap_or instead of or_else().unwrap()

* weather: Move data into modules

met_no needs a one-time fetch of the "legend", the translation table
from weather summary keywords to textual representations. This is now
stored in the Config struct within the module, and fetched on first get.

Adds a trait WeatherProvider with a single function get_weather that
takes a mutable self, to allow for this update.

The main loop instantiates the config within an Arc<Mutex>>. Async
FnMut enclosures cannot have references that outlive the enclosure, so
to use api.recoverable we need to do an Arc clone and lock. This also
makes sure only one request can be run at a time.

Also moves the autolocation call inside the api.recoverable block.

* weather: Reinstate apparent temp calculation

Reverting the change in #1573.

This version is metric units only, as it is only needed for met_no.

* weather: Propagate error in find_ip_location()

* weather: Move initialization to new()

Avoid mutating the config object by creating a wrapper Service object
and instantiate that with new() on block initialization. This also means
we can remove Mutex and Arc, and instead just keep the Service in a Box.

* doc(weather): Met_no now supports apparent temp

* weather(met_no): Run get_legend in api.recoverable

* remove unnecessary `pub`s

* remove some clones

Co-authored-by: Sigve Indregard <sigve.indregard@pm.me>
Co-authored-by: Max Verevkin <maxxverrr@gmail.com>