Add `not_charging_format` config option
It is possible for a battery to be in the "Not charging" state, in
multi-battery setups especially, where both batteries may be partially
drained, but only one will be being charged at once. In such cases the
battery block incorrectly reports the non-charging battery as full.

In order to not break existing behaviour, and also given as mentioned in
continue to report "Not charging" batteries to be full by default.

Fixes #1416.