blocks/watson: Avoid error on race condition while reading state (#1776)
While reading state, neither we nor watson seem to perfom
locking of the state file. Thus, parsing the state file can fail
unexpectedly. This would lead to an error in the block, making
it unusable until i3status-rust is restartet. This change
introduces the empty state as a default to avoid this behavior.

fixes #1775