Stop find_files requiring a specific extension (#1779)
If a file with no extension is passed as the config file, the error
saying it doesn't exist is misleading. Similarly, if a config.cfg is
passed in, it's surprising for the user that config.toml is the file
that's used.

This changes it to leave any existing extension as is but to also check
for existence if the given extension (if any) is added.