warn clippy::unused_trait_names
Checks for use Trait where the Trait is only used for its methods and not referenced by a path directly.

Traits imported that aren’t used directly can be imported anonymously with use Trait as _. It is more explicit, avoids polluting the current scope with unused names and can be useful to show which imports are required for traits.