disk_space: Support btrfs backend
Btrfs is too smart for the statvfs based backend (i.e., only counting
blocks leads to wrong numbers).

For example, a btrfs disk with a lot of de-duplicated blocks (via the copy
on write mechanism) might have a drastically over-reported disk usage.

The btrfs backend is currently implemented by parsing the output of the
`btrfs filesystem usage --raw` command. This is suboptimal, as this now
relies on the command output not changing.

Vendoring the algorithm used internally by the `btrfs` command does not
seem to be a reasonable alternative, considering that the code[1] is
rather complex, low level and would require semi-constant maintenance.
Additionally, the c code would need bindings to be usable from rust.

I assume, that the `btrfs` command output will stay rather similar in
the future, as a lot of tools rely on directly parsing it (see the
various scripts in the issue, this commit fixes).

[1]: https://github.com/kdave/btrfs-progs/blob/eeab081e9d9fbdf4583122ed1caedf541383cf2d/cmds/filesystem-usage.c#L442

Fixes: #1654