blocks[battery[apc_ups]]: Deserialize APC UPS data
Created a new `Value` enum to represent the different types of values
that are used.

Created a new `Properties` struct to represent the properties of the
APC UPS data.

Use `Framed` with `LengthDelimitedCodec` to decode the frames of data,
before deserializing them into the `Properties` struct.

Added the ability to decode not only the time remaining in minutes, but
also seconds or hours.