fix(alsa): Don't leak alsactl process
This bug was found when alsactl was repeatedly failing due to a missing
device causing a number of alsactl processes to be running.

This PR uses tokio's kill_on_drop and an added the process handle to a field
in the device. I have not checked the rest of the code base for similar
issues but at least after this PR alsa is fixed. :)